pyproject.toml
[tool.poetry]
name = "fastapi_practice"
version = "0.1.0"
description = ""
authors = ["hhhscvx <hhhscvx@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.32"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
PyJWT = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
